requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pypdf==3.17.4
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.0
reportlab==4.0.9
python-dateutil==2.8.2
jinja2==3.1.3

# Packaging / Desktop
pywebview==3.7
pyinstaller==5.16.0
